Move serializer to internal
