+++ src/internal/serializer/serialize.ts
@@ -1,7 +1,7 @@
-import { isUndefined } from '../internal/typeChecking';
-import { QueryParams, RawQueryParams, IQueryParamsSchema } from '../types';
+import { isUndefined } from '../typeChecking';
+import { QueryParams, RawQueryParams, IQueryParamsSchema } from '../../types';
 
-import { Errors } from '../errors';
+import { Errors } from '../../errors';
 
 /**
  * Parse each individual query params with the parser provided for each prop
